refactor CSRequest to use EducatorProfile handler
