--- db/migrate/20220127192512_add_sheerid_webhook_received_to_users.rb
@@ -1,5 +1,0 @@
-class AddSheeridWebhookReceivedToUsers < ActiveRecord::Migration[5.2]
-  def change
-    add_column :users, :sheer_id_webhook_received, :boolean
-  end
-end
